requirements.txt
requests 
beautifulsoup4 
python-docx 
streamlit 